added training loop

+++ masks.py
@@ -10,7 +10,7 @@ def mask1(input_output_size,num_layers):
      [0,0,0,1,1,1]
     ]
     """
-    return torch.randint(2,(num_layers,input_output_size))
+    return torch.randint(2,(num_layers,input_output_size)).float()
 
 def mask2(input_output_size,num_layers):
     """
@@ -20,4 +20,4 @@ def mask2(input_output_size,num_layers):
      [0,1,0,1,0,1]
     ]
     """
-    return torch.tensor([[(i+j)%2 for i in range(input_output_size)] for j in range(num_layers)])
+    return torch.tensor([[(i+j)%2 for i in range(input_output_size)] for j in range(num_layers)]).float()
